Restructure source
Fixes issue #1.
